corrections in models and migrations

+++ db/migrate/20210407190714_create_dishes.rb
@@ -3,7 +3,7 @@ class CreateDishes < ActiveRecord::Migration[6.1]
     create_table :dishes do |t|
       t.string :title, null: false
       t.text :description
-      t.integer :price, null: false
+      t.float :price, null: false
       t.references :category, null: false, foreign_key: true
 
       t.timestamps
